1. Debug statements added in Resize.cpp - Needs clean up 2. Debug statements added in node_resize.cpp -Needs clean up 3. The dtype changes from size_t to uint64_t in tensor.cpp & image_read_and_decode.cpp 4. Reduced the max_dec_w/h to 1000x1000 for image_slice 5. Debug statements added in pytorch.py - Needs clean up

+++ rocAL/rocAL/source/loaders/image/image_read_and_decode.cpp
@@ -148,7 +148,7 @@ ImageReadAndDecode::load(unsigned char* buff,
     const Decoder::ColorFormat decoder_color_format = std::get<0>(ret);
     const unsigned output_planes = std::get<1>(ret);
     const bool keep_original = decoder_keep_original;
-    const size_t image_size = max_decoded_width * max_decoded_height * output_planes * sizeof(unsigned char);
+    const uint64_t image_size = max_decoded_width * max_decoded_height * output_planes * sizeof(unsigned char);
 
     // Decode with the height and size equal to a single image
     // File read is done serially since I/O parallelization does not work very well.
@@ -157,7 +157,7 @@ ImageReadAndDecode::load(unsigned char* buff,
         while ((file_counter != _batch_size) && _reader->count_items() > 0)
         {
             auto read_ptr = buff + image_size * file_counter;
-            size_t fsize = _reader->open();
+            uint64_t fsize = _reader->open();
             if (fsize == 0) {
                 WRN("Opened file " + _reader->id() + " of size 0");
                 continue;
@@ -183,7 +183,7 @@ ImageReadAndDecode::load(unsigned char* buff,
     else {
         while ((file_counter != _batch_size) && _reader->count_items() > 0) {
 
-            size_t fsize = _reader->open();
+            uint64_t fsize = _reader->open();
             if (fsize == 0) {
                 WRN("Opened file " + _reader->id() + " of size 0");
                 continue;
